added:  - hack to track first db creation to not remove contacts from server because db is always empty on first launch

+++ app/src/main/java/com/iyakovlev/contacts/data/repository/contacts/ContactsRepositoryImpl.kt
@@ -1,5 +1,7 @@
 package com.iyakovlev.contacts.data.repository.contacts
 
+import androidx.room.RoomDatabase
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.iyakovlev.contacts.R
 import com.iyakovlev.contacts.common.constants.Constants
 import com.iyakovlev.contacts.common.constants.Constants.ISDEBUG
@@ -8,6 +10,7 @@ import com.iyakovlev.contacts.data.api.ApiService
 import com.iyakovlev.contacts.data.database.repository.DatabaseRepository
 import com.iyakovlev.contacts.data.repository.user.UserRepositoryImpl
 import com.iyakovlev.contacts.domain.model.UserRemote
+import com.iyakovlev.contacts.utils.AppStatus.isFirstLaunch
 import com.iyakovlev.contacts.utils.log
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -70,21 +73,26 @@ class ContactsRepositoryImpl @Inject constructor(
             onSuccess = {// TODO: maybe refactor somehow
                 // makes server and db info the same:
 
-                // add contacts that are in db but not on server
-                val l = db.getContacts().filter { contactEntity ->
-                    !it.data.contacts.map { it.toContactEntity() }.contains(contactEntity)
-                }
-                log("difference: $l", ISDEBUG)
-                for (i in l) {
-                    addContact(i.id)
-                }
+                val a = db.getContacts()
+                if (!isFirstLaunch) { // todo: maybe delete this hack
+                    // TODO: temporary table with deleted contacts maybe
+                    // add contacts that are in db but not on server
+                    val l = db.getContacts().filter { contactEntity ->
+                        !it.data.contacts.map { it.toContactEntity() }.contains(contactEntity)
+                    }
+                    log("difference: $l", ISDEBUG)
+                    for (i in l) {
+                        addContact(i.id)
+                    }
 
-                // delete contacts that are deleted from db but not deleted from server
-                val ll = it.data.contacts.map { it.toUserRemote() }.filter { user ->
-                    !db.getContacts().map { it.toUserRemote() }.contains(user)
-                }
-                for (i in ll) {
-                    deleteContact(i.id)
+                    // TODO: check first launch
+                    // delete contacts that are deleted from db but not deleted from server
+                    val ll = it.data.contacts.map { it.toUserRemote() }.filter { user ->
+                        !db.getContacts().map { it.toUserRemote() }.contains(user)
+                    }
+                    for (i in ll) {
+                        deleteContact(i.id)
+                    }
                 }
 
                 // return list from server
